+++ benchmark/src/libfork/fib.cpp
@@ -1,8 +1,8 @@
 #include <benchmark/benchmark.h>
 
-#include "libfork_benchmark/common.hpp"
+#include "common.hpp"
 
-#include "libfork_benchmark/fib/fib.hpp"
+#include "fib.hpp"
 
 import std;
